pyproject.toml
[project]
name = "bix-plaza"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "William Fernandes Dias", email = "william.winchester1967@gmail.com" },
]
dependencies = [
    "Django==4.2.*",
    "Celery>=5.3.6",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.1",
    "python-decouple>=3.8",
    "redis>=5.0.3",
    "dj-database-url>=2.1.0",
    "psycopg[binary]>=3.1.18",
    "django-extensions>=3.2.3",
    "django-filter>=24.1",
    "djangorestframework>=3.14.0",
    "django-phonenumber-field[phonenumbers]>=7.3.0",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-access-policy>=1.5.0",
    "drf-extensions>=0.7.1",
    "django-anymail[sendgrid]>=10.3",
    "django-fsm>=2.8.1",
    "setuptools>=69.2.0",
    "pytest==7.1.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.7.1",
    "pytest<8,>=7.1.3",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.8.0",
    "Faker>=24.2.0",
    "factory-boy>=3.3.0",
    "pytest-mock>=3.12.0",
    "watchfiles>=0.21.0",
    "django-stubs[compatible-mypy]>=4.2.7",
    "djangorestframework-stubs[compatible-mypy]>=3.14.5",
    "ruff>=0.3.3",
    "django-debug-toolbar>=4.3.0",
    "pytest-xdist>=3.5.0",
    "pytest-drf>=1.1.3",
    "pytest-camel-collect>=1.0.2",
    "sqlalchemy>=2.0.28",
]

[tool.pdm.scripts]
"test" = "pytest --cov=src"
"test:watch" = "watchfiles 'pytest --cov=src' src tests"


[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING",
    "if __name__ == .__main__.:"
]


[tool.pytest.ini_options]
addopts = "--no-migrations --no-cov-on-fail"
filterwarnings = ["error", "ignore::DeprecationWarning"]
pythonpath = ["src", "."]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_classes = ["Test-*", "Describe-*", "Context-*", "With-*", "Without-*", "For-*", "When-*", "If-*", "Case-*"]
python_functions = ["test_*", "it_*", "its_*"]

[tool.ruff]
line-length = 88
indent-width = 4
src = ["src", "tests"]
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
docstring-code-line-length = 120
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false

[tool.ruff.lint]
fixable = ["ALL"]
select = ["C", "E", "F", "W", "B9", "D", "DJ"]
ignore = ["E203", "E266", "D100", "D203", "D213", "D104", "D106", "D101", "D102"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501"]
"asgi.py" = ["D212"]

